Ramp.MNT: move auxiliar GraphKicks widget to custom_widgets module
