+++ pyqt-apps/siriushla/bo_ap_ramp/custom_widgets.py
@@ -1,7 +1,12 @@
 
+import numpy as _np
+
+from qtpy.QtGui import QColor
 from qtpy.QtCore import Qt, QLocale
 from qtpy.QtWidgets import QLineEdit, QTableWidget, QTableWidgetItem, \
-    QStyledItemDelegate, QDoubleSpinBox
+    QStyledItemDelegate, QDoubleSpinBox, QToolTip
+from pyqtgraph import functions
+from pydm.widgets import PyDMWaveformPlot
 
 from siriushla.as_ap_configdb import LoadConfigDialog as _LoadConfigDialog
 
@@ -102,3 +107,53 @@ class ConfigLineEdit(QLineEdit):
         popup = _LoadConfigDialog('bo_normalized')
         popup.configname.connect(self.setText)
         popup.exec_()
+
+
+class GraphKicks(PyDMWaveformPlot):
+
+    def __init__(self, parent=None, xdata=list(), ydata=list(),
+                 tooltip_names=list(), c0=0, color='blue'):
+        super().__init__(parent)
+        self.setBackgroundColor(QColor(255, 255, 255))
+        self.setAutoRangeX(True)
+        self.setAutoRangeY(True)
+        self.setShowXGrid(True)
+        self.setShowYGrid(True)
+        self.setObjectName('graph')
+        self.setStyleSheet('#graph{min-width:32em;min-height:14em;}')
+
+        self.xdata = xdata
+        self.ydata = ydata
+        self.tooltip_names = tooltip_names
+        self.c0 = c0
+
+        self.addChannel(
+            y_channel='Kicks', x_channel='Pos',
+            color=color, lineWidth=2, symbol='o', symbolSize=10)
+        self.curve = self.curveAtIndex(0)
+        self.curve.receiveXWaveform(xdata)
+        self.curve.receiveYWaveform(ydata)
+        self.curve.redrawCurve()
+
+        self.addChannel(
+            y_channel='Mean', x_channel='Pos',
+            color='black', lineStyle=1, lineWidth=2)
+        self.mean = self.curveAtIndex(1)
+        self.mean.receiveXWaveform(xdata)
+        self.mean.receiveYWaveform(_np.array([_np.mean(ydata)]*len(ydata)))
+        self.mean.redrawCurve()
+
+    def mouseMoveEvent(self, ev):
+        unit = 'urad'
+        pos = ev.pos()
+
+        posx = self.curve.scatter.mapFromScene(pos).x()
+        posx = posx % self.c0
+        ind = _np.argmin(_np.abs(_np.array(self.xdata)-posx))
+        posy = self.curve.scatter.mapFromScene(pos).y()
+
+        sca, prf = functions.siScale(posy)
+        txt = '{0:s}, y = {1:.3f} {2:s}'.format(
+            self.tooltip_names[ind], sca*posy, prf+unit)
+        QToolTip.showText(
+            self.mapToGlobal(pos), txt, self, self.geometry(), 500)
